+++ src/modules/projects/error.ts
@@ -3,3 +3,9 @@ export class ProjectNotFound extends Error {
     super("Project not found");
   }
 }
+
+export class ProjectInvalidPosition extends Error {
+  constructor() {
+    super("Either afterId or beforeId must be provided and they can't be the same");
+  }
+}